Transforms OverlaySection into  external component
