+++ src/pages/programs/index.tsx
@@ -5,6 +5,7 @@ import Row from 'react-bootstrap/Row';
 
 import { Bar } from '../../components/bar';
 import { CourseCard } from '../../components/course-card';
+import OverlaySection from '../../components/overlay-section';
 import { DefaultLayout } from '../../layouts/default-layout';
 
 import HeroHome from '../../images/backgrounds/hero-home.jpg';
@@ -105,7 +106,7 @@ const ProgramsPage: NextPage = () => (
     <OverlaySection backgroundImage={EventWedding}>
       <>
         <h2>QC Event School</h2>
-        <p className="lead">The event planning industry is expecting a 110% growth over the next 10 years! If you’re organized, detail-oriented, and enjoy working with people, you might just have found your new career in event and wedding planning!  QC provides fundamental, specialized, and advanced courses for event planning professionals of all skill levels.</p>
+        <p className="lead">The event planning industry is expecting a 110% growth over the next 10 years! If you’re organized, detail-oriented, and enjoy working with people, you might just have found your new career in event and wedding planning! QC provides fundamental, specialized, and advanced courses for event planning professionals of all skill levels.</p>
       </>
     </OverlaySection>
 
@@ -341,30 +342,9 @@ const ProgramsPage: NextPage = () => (
         background-size: cover;
       }
     `}</style>
-
-  </DefaultLayout >
+  </DefaultLayout>
 );
 
 const CardColumn: React.FC = ({ children }) => <div className="col-12 col-md-8 col-lg-6 offset-md-2 offset-lg-0 mb-g d-flex">{children}</div>;
-
-const OverlaySection: React.FC<{ backgroundImage: string; }> = ({ children, backgroundImage }) => (
-  <section className="overlay-section bg-dark text-light text-shadow text-right">
-    <Container>
-      <Row>
-        <Col xs={12} md={8} className="offset-md-4">
-          <Bar className="text-primary ml-auto" />
-          {children}
-        </Col>
-      </Row>
-    </Container>
-
-    <style jsx>{`
-      section {
-        background: url(${backgroundImage}) 0 0;
-        background-size: cover;
-      }
-    `}</style>
-  </section>
-);
 
 export default ProgramsPage;
